# apps/web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Action-Output": {
        "additionalProperties": false,
        "description": "Proposed mitigation action with impact and parameters.",
        "properties": {
          "confidence": {
            "$ref": "#/components/schemas/Confidence-Output",
            "description": "Confidence in the action's effectiveness"
          },
          "description": {
            "$ref": "#/components/schemas/Sentence-Output",
            "description": "Detailed action description with impact figures"
          },
          "id": {
            "description": "Unique action ID",
            "example": "action_raise_price_3pct",
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Impact"
              },
              {
                "type": "null"
              }
            ],
            "description": "Expected business impact if executed"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "description": "Editable parameters (price, qty, etc.)",
            "title": "Parameters",
            "type": "object"
          },
          "title": {
            "description": "Short action title",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Type: email_draft, task, purchase_order_draft, or price_change_draft",
            "enum": [
              "email_draft",
              "task",
              "purchase_order_draft",
              "price_change_draft"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "type",
          "impact",
          "confidence",
          "parameters"
        ],
        "title": "Action",
        "type": "object"
      },
...
      "Alert": {
        "additionalProperties": false,
        "description": "Complete alert lifecycle state.",
        "properties": {
          "actions": {
            "description": "Proposed mitigation actions",
            "items": {
              "$ref": "#/components/schemas/Action-Output"
            },
            "maxItems": 3,
            "minItems": 0,
            "title": "Actions",
            "type": "array"
          },
          "canDecide": {
            "default": false,
            "description": "Whether the person signed in may decide this alert, computed per request and never stored",
            "title": "Candecide",
            "type": "boolean"
          },
          "cause": {
            "description": "Root cause (identified with evidence or no_evidence)",
            "discriminator": {
              "mapping": {
                "identified": "#/components/schemas/CauseIdentified-Output",
                "no_evidence": "#/components/schemas/CauseNoEvidence-Output"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CauseIdentified-Output"
              },
              {
                "$ref": "#/components/schemas/CauseNoEvidence-Output"
              }
            ],
            "title": "Cause"
          },
          "changesRequested": {
            "default": false,
            "description": "Whether a person asked for changes to the proposal",
            "title": "Changesrequested",
            "type": "boolean"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence-Output",
            "description": "Confidence in the alert detection"
          },
          "decidedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who decides this alert: the area that owns its metric, or Gerencia when no area owns it",
            "title": "Decidedby"
          },
          "executedAction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutedAction"
              },
              {
                "type": "null"
              }
            ],
            "description": "Action executed (only when status=executed)"
          },
          "id": {
            "description": "Unique alert ID",
            "example": "alerta_abc123def456",
            "title": "Id",
            "type": "string"
          },
          "labels": {
            "description": "Identity labels a person reads, the metric's short name first then the affected entity (e.g. ['Margen', 'l\u00ednea Hogar'])",
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "mergedAlerts": {
            "description": "Alerts merged into this one",
            "items": {
              "$ref": "#/components/schemas/MergedAlert"
            },
            "title": "Mergedalerts",
            "type": "array"
          },
          "mergedInto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ID of the alert this one was merged into",
            "title": "Mergedinto"
          },
          "metric": {
            "description": "Metric that triggered this alert",
            "enum": [
              "margen_pct",
              "saldo_vencido",
              "dias_pago_prom",
              "cobertura_dias",
              "descuento_en_exceso",
              "veces_intervalo_habitual"
            ],
            "title": "Metric",
            "type": "string"
          },
          "pesosAtRisk": {
            "$ref": "#/components/schemas/Figure",
            "description": "Pesos at risk (COP)"
          },
          "recoverablePerMonth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Figure"
              },
              {
                "type": "null"
              }
            ],
            "description": "Potential monthly recovery if action taken"
          },
          "severity": {
            "description": "Alert severity: critical, high, medium, or low",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ],
            "title": "Severity",
            "type": "string"
          },
          "simulatedDate": {
            "description": "Simulated date when alert was created (ISO 8601)",
            "title": "Simulateddate",
            "type": "string"
          },
          "status": {
            "description": "Current status in lifecycle: new \u2192 analyzing \u2192 proposed \u2192 approved/rejected \u2192 executed",
            "enum": [
              "new",
              "analyzing",
              "proposed",
              "approved",
              "rejected",
              "executed",
              "merged"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/Sentence-Output",
            "description": "Alert title with initial impact figures"
          }
        },
        "required": [
          "id",
          "status",
          "severity",
          "metric",
          "labels",
          "title",
          "pesosAtRisk",
          "recoverablePerMonth",
          "confidence",
          "simulatedDate",
          "cause",
          "actions",
          "executedAction",
          "changesRequested",
          "mergedInto",
          "mergedAlerts",
          "decidedBy",
          "canDecide"
        ],
        "title": "Alert",
        "type": "object"
      },
      "AlertEstadoEnum": {
        "description": "Valid alert estado values (Spanish names for API contract).",
        "enum": [
          "nueva",
          "en_analisis",
          "propuesta",
          "aprobada",
          "rechazada",
          "ejecutada"
        ],
        "title": "AlertEstadoEnum",
        "type": "string"
      },
...
      "CauseIdentified-Output": {
        "additionalProperties": false,
        "description": "Root cause found with evidence.",
        "properties": {
          "evidence": {
            "description": "Supporting evidence",
            "items": {
              "$ref": "#/components/schemas/Evidence-Output"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "kind": {
            "const": "identified",
            "default": "identified",
            "description": "Discriminator: cause was identified",
            "title": "Kind",
            "type": "string"
          },
          "sentence": {
            "$ref": "#/components/schemas/Sentence-Output",
            "description": "Root cause statement with figures"
          }
        },
        "required": [
          "kind",
          "sentence",
          "evidence"
        ],
        "title": "CauseIdentified",
        "type": "object"
      },
...
      "CauseNoEvidence-Output": {
        "additionalProperties": false,
        "description": "Root cause could not be identified with available data.",
        "properties": {
          "kind": {
            "const": "no_evidence",
            "default": "no_evidence",
            "description": "Discriminator: cause could not be identified",
            "title": "Kind",
            "type": "string"
          },
          "queriesReviewed": {
            "description": "Query IDs that were searched",
            "items": {
              "type": "string"
            },
            "title": "Queriesreviewed",
            "type": "array"
          },
          "reason": {
            "description": "Explanation why cause could not be determined",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "reason",
          "queriesReviewed"
        ],
        "title": "CauseNoEvidence",
        "type": "object"
      },
...
      "Confidence-Output": {
        "additionalProperties": false,
        "description": "Confidence assessment with assumptions.",
        "properties": {
          "assumptions": {
            "description": "Assumptions underlying the confidence assessment",
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "level": {
            "description": "Confidence level: high, medium, or low",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "level",
          "assumptions"
        ],
        "title": "Confidence",
        "type": "object"
      },
...
      "DecisionApprove": {
        "additionalProperties": false,
        "description": "Human decision: approve and execute action.",
        "properties": {
          "actionId": {
            "description": "ID of action to approve",
            "title": "Actionid",
            "type": "string"
          },
          "kind": {
            "const": "approve",
            "default": "approve",
            "description": "Discriminator: approve decision",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "actionId"
        ],
        "title": "DecisionApprove",
        "type": "object"
      },
      "DecisionEdit": {
        "additionalProperties": false,
        "description": "Human decision: approve with modified parameters.",
        "properties": {
          "actionId": {
            "description": "ID of action to modify and approve",
            "title": "Actionid",
            "type": "string"
          },
          "kind": {
            "const": "edit",
            "default": "edit",
            "description": "Discriminator: edit and approve decision",
            "title": "Kind",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "description": "Modified parameters (price, qty, etc.)",
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "actionId",
          "parameters"
        ],
        "title": "DecisionEdit",
        "type": "object"
      },
      "DecisionReject": {
        "additionalProperties": false,
        "description": "Human decision: reject alert.",
        "properties": {
          "kind": {
            "const": "reject",
            "default": "reject",
            "description": "Discriminator: reject decision",
            "title": "Kind",
            "type": "string"
          },
          "reason": {
            "description": "Reason for rejection",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DecisionReject",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Query": {
        "additionalProperties": false,
        "description": "A query of the kernel that produced a figure, recorded when an agent ran it.",
        "properties": {
          "description": {
            "description": "The KPI and the simulated day it was read on",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "description": "The figure's queryId",
            "title": "Id",
            "type": "string"
          },
          "source": {
            "default": "kernel",
            "description": "Where the query runs: the KPI kernel of packages/tools, or the stored alerts",
            "enum": [
              "kernel",
              "alertas"
            ],
            "title": "Source",
            "type": "string"
          },
          "sql": {
            "description": "The call the kernel ran",
            "title": "Sql",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "sql",
          "description"
        ],
        "title": "Query",
        "type": "object"
      },
...
      "SimulatedDay": {
        "additionalProperties": false,
        "description": "Current simulated day in ISO 8601 format.",
        "properties": {
          "dia": {
            "description": "Current simulated day (YYYY-MM-DD)",
            "example": "2026-10-03",
            "title": "Dia",
            "type": "string"
          }
        },
        "required": [
          "dia"
        ],
        "title": "SimulatedDay",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Centinela API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/alertas": {
      "get": {
        "description": "List all alerts, optionally filtered by estado.\n\nValid valores: nueva, en_analisis, propuesta, aprobada, rechazada, ejecutada",
        "operationId": "listar_alertas_get",
        "parameters": [
          {
            "description": "Filter by alert estado (Spanish name for status)",
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlertEstadoEnum"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by alert estado (Spanish name for status)",
              "title": "Estado"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  },
                  "title": "Response Listar Alertas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar",
        "tags": [
          "alerts"
        ]
      }
    },
    "/alertas/{id}": {
      "get": {
        "operationId": "obtener_alertas__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obtener",
        "tags": [
          "alerts"
        ]
      }
    },
    "/alertas/{id}/decision": {
      "post": {
        "operationId": "decidir_alertas__id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "approve": "#/components/schemas/DecisionApprove",
                    "edit": "#/components/schemas/DecisionEdit",
                    "reject": "#/components/schemas/DecisionReject"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DecisionApprove"
                  },
                  {
                    "$ref": "#/components/schemas/DecisionEdit"
                  },
                  {
                    "$ref": "#/components/schemas/DecisionReject"
                  }
                ],
                "title": "Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decidir",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/consultas/{query_id}": {
      "get": {
        "description": "The kernel query that produced a figure, as the agent that cited it recorded it.",
        "operationId": "obtener_consultas__query_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "title": "Query Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Query"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obtener",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/simulacion/avanzar": {
      "post": {
        "operationId": "avanzar_simulacion_avanzar_post",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Avanzar"
      }
    },
    "/simulacion/dia-actual": {
      "get": {
        "description": "Get the current simulated day.\nUsed by Vig\u00eda to know what date to filter queries by.",
        "operationId": "dia_actual_simulacion_dia_actual_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatedDay"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dia Actual"
      }
    }
  }
}